requirements.txt
black
isort
pytest
ruff
python-dotenv
pytest-cov
tabulate